reduce_cpp: use vector iterator to compute sum
This commit makes the computed sum match what is now assumed to be the
correct value by making the value range from 0 to LIMIT inclusive.

Additionally, this commit updates the way the vector of integers is
handled. Previously, values were assigned to the vector as if it were a
plain array. The values are now assigned via push_back() so that the
vector's size is correct and an iterator created from the vector can
actually iterate over the container.